backend/requirements.txt
amqp==5.0.6
bidict==0.21.2
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SocketIO==5.1.0
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
h11==0.12.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
praw==7.3.0
prawcore==2.2.0
prompt-toolkit==3.0.19
python-dotenv==0.18.0
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
simple-websocket==0.2.0
six==1.16.0
update-checker==0.18.0
urllib3==1.26.6
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.1.0
Werkzeug==2.0.1
wsproto==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
